pyright: Fulfill strict typing requirements

+++ cobbler/actions/sync.py
@@ -11,15 +11,16 @@ import glob
 import logging
 import os
 import time
-from typing import TYPE_CHECKING, List, Optional
+from typing import TYPE_CHECKING, Dict, List, Optional
 
 from cobbler import templar, tftpgen, utils
 from cobbler.cexceptions import CX
 from cobbler.utils import filesystem_helpers
 
 if TYPE_CHECKING:
+    from cobbler.api import CobblerAPI
     from cobbler.items.profile import Profile
-    from cobbler.items.system import System
+    from cobbler.manager import ManagerModule
 
 
 class CobblerSync:
@@ -27,7 +28,14 @@ class CobblerSync:
     Handles conversion of internal state to the tftpboot tree layout
     """
 
-    def __init__(self, api, verbose: bool = True, dhcp=None, dns=None, tftpd=None):
+    def __init__(
+        self,
+        api: "CobblerAPI",
+        verbose: bool = True,
+        dhcp: Optional["ManagerModule"] = None,
+        dns: Optional["ManagerModule"] = None,
+        tftpd: Optional["ManagerModule"] = None,
+    ) -> None:
         """
         Constructor
 
@@ -49,8 +57,14 @@ class CobblerSync:
         self.repos = api.repos()
         self.templar = templar.Templar(self.api)
         self.tftpgen = tftpgen.TFTPGen(api)
+        if dns is None:
+            raise ValueError("dns not optional")
         self.dns = dns
+        if dhcp is None:
+            raise ValueError("dns not optional")
         self.dhcp = dhcp
+        if tftpd is None:
+            raise ValueError("dns not optional")
         self.tftpd = tftpd
         self.bootloc = self.settings.tftpboot_location
 
@@ -85,20 +99,20 @@ class CobblerSync:
         self.settings = self.api.settings()
         self.repos = self.api.repos()
 
-    def run_sync_systems(self, systems: List["System"]):
+    def run_sync_systems(self, systems: List[str]):
         """
         Syncs the specific systems with the config tree.
         """
         self.__common_run()
 
         # Have the tftpd module handle copying bootloaders, distros, images, and all_system_files
-        self.tftpd.sync_systems(systems)
+        self.tftpd.sync_systems(systems)  # type: ignore
 
         if self.settings.manage_dhcp:
             self.write_dhcp()
         if self.settings.manage_dns:
             self.logger.info("rendering DNS files")
-            self.dns.regen_hosts()
+            self.dns.regen_hosts()  # type: ignore
             self.dns.write_configs()
 
         self.logger.info("cleaning link caches")
@@ -113,7 +127,7 @@ class CobblerSync:
         utils.run_triggers(self.api, None, "/var/lib/cobbler/triggers/sync/post/*")
         utils.run_triggers(self.api, None, "/var/lib/cobbler/triggers/change/*")
 
-    def run(self):
+    def run(self) -> None:
         """
         Syncs the current configuration file with the config tree.
         Using the ``Check().run_`` functions previously is recommended
@@ -125,7 +139,7 @@ class CobblerSync:
         self.clean_trees()
 
         # Have the tftpd module handle copying bootloaders, distros, images, and all_system_files
-        self.tftpd.sync(self.verbose)
+        self.tftpd.sync()
         # Copy distros to the webdir
         # Adding in the exception handling to not blow up if files have been moved (or the path references an NFS
         # directory that's no longer mounted)
@@ -143,12 +157,12 @@ class CobblerSync:
             self.write_dhcp()
         if self.settings.manage_dns:
             self.logger.info("rendering DNS files")
-            self.dns.regen_hosts()
+            self.dns.regen_hosts()  # type: ignore
             self.dns.write_configs()
 
         if self.settings.manage_tftpd:
             # copy in boot_files
-            self.tftpd.write_boot_files()
+            self.tftpd.write_boot_files()  # type: ignore
 
         self.logger.info("cleaning link caches")
         self.clean_link_cache()
@@ -219,7 +233,7 @@ class CobblerSync:
         """
         if self.settings.manage_dhcp:
             self.write_dhcp()
-            self.dhcp.sync_dhcp()
+            self.dhcp.sync_dhcp()  # type: ignore
 
     def clean_link_cache(self):
         """
@@ -245,7 +259,7 @@ class CobblerSync:
                 ]
                 utils.subprocess_call(cmd, shell=False)
 
-    def rsync_gen(self):
+    def rsync_gen(self) -> None:
         """
         Generate rsync modules of all repositories and distributions
 
@@ -259,7 +273,7 @@ class CobblerSync:
         except Exception as error:
             raise OSError(f"error reading template {template_file}") from error
 
-        distros = []
+        distros: List[Dict[str, str]] = []
 
         for link in glob.glob(os.path.join(self.settings.webdir, "links", "*")):
             distro = {}
@@ -285,7 +299,7 @@ class CobblerSync:
 
         self.templar.render(template_data, metadata, "/etc/rsyncd.conf")
 
-    def add_single_distro(self, name):
+    def add_single_distro(self, name: str) -> None:
         """
         Sync adding a single distro.
 
@@ -293,11 +307,11 @@ class CobblerSync:
         """
         # get the distro record
         distro = self.distros.find(name=name)
-        if distro is None:
+        if distro is None or isinstance(distro, list):
             return
         # copy image files to images/$name in webdir & tftpboot:
         self.tftpgen.copy_single_distro_files(distro, self.settings.webdir, True)
-        self.tftpd.add_single_distro(distro)
+        self.tftpd.add_single_distro(distro)  # type: ignore
 
         # create the symlink for this distro
         src_dir = distro.find_distro_path()
@@ -325,25 +339,31 @@ class CobblerSync:
         # generate any templates listed in the distro
         self.tftpgen.write_templates(distro, write_file=True)
         # cascade sync
-        kids = self.api.find_items("profile", {"distro": name}, return_list=True)
+        kids = self.api.find_profile(return_list=True, **{"distro": name})
+        if not isinstance(kids, list):
+            raise ValueError("Expected to get list of profiles from search!")
         for k in kids:
             self.add_single_profile(k, rebuild_menu=False)
         self.tftpgen.make_pxe_menu()
 
-    def add_single_image(self, name):
+    def add_single_image(self, name: str) -> None:
         """
         Sync adding a single image.
 
         :param name: The name of the image.
         """
         image = self.images.find(name=name)
+        if image is None or isinstance(image, list):
+            return
         self.tftpgen.copy_single_image_files(image)
-        kids = self.api.find_items("system", {"image": name})
+        kids = self.api.find_system(return_list=True, **{"image": name})
+        if not isinstance(kids, list):
+            raise ValueError("Expected to get list of profiles from search!")
         for k in kids:
-            self.add_single_system(k)
+            self.add_single_system(k.name)
         self.tftpgen.make_pxe_menu()
 
-    def remove_single_distro(self, name):
+    def remove_single_distro(self, name: str) -> None:
         """
         Sync removing a single distro.
 
@@ -362,7 +382,7 @@ class CobblerSync:
             name + "*.repo",
         )
 
-    def remove_single_image(self, name):
+    def remove_single_image(self, name: str) -> None:
         """
         Sync removing a single image.
 
@@ -381,25 +401,26 @@ class CobblerSync:
         :param rebuild_menu: Whether to rebuild the grub/... menu or not.
         :return: ``True`` if this succeeded.
         """
-        # get the profile object:
-        if profile is None:
+        if profile is None or isinstance(profile, list):  # type: ignore
             # Most likely a subprofile's kid has been removed already, though the object tree has not been reloaded and
             # this is just noise.
-            return
+            return None
         # Rebuild the yum configuration files for any attached repos generate any templates listed in the distro.
         self.tftpgen.write_templates(profile)
         # Cascade sync
         kids = profile.children
         for k in kids:
-            self.add_single_profile(k, rebuild_menu=False)
-        kids = self.api.find_items("system", {"profile": profile.name})
+            self.add_single_profile(k, rebuild_menu=False)  # type: ignore
+        kids = self.api.find_system(return_list=True, **{"profile": profile.name})
+        if not isinstance(kids, list):
+            raise ValueError("Expected to get list of profiles from search!")
         for k in kids:
-            self.add_single_system(k)
+            self.add_single_system(k.name)
         if rebuild_menu:
             self.tftpgen.make_pxe_menu()
         return True
 
-    def remove_single_profile(self, name: str, rebuild_menu: bool = True):
+    def remove_single_profile(self, name: str, rebuild_menu: bool = True) -> None:
         """
         Sync removing a single profile.
 
@@ -415,7 +436,7 @@ class CobblerSync:
         if rebuild_menu:
             self.tftpgen.make_pxe_menu()
 
-    def update_system_netboot_status(self, name: str):
+    def update_system_netboot_status(self, name: str) -> None:
         """
         Update the netboot status of a system.
 
@@ -424,9 +445,9 @@ class CobblerSync:
         system = self.systems.find(name=name)
         if system is None:
             return
-        self.tftpd.sync_single_system(system)
+        self.tftpd.sync_single_system(system)  # type: ignore
 
-    def add_single_system(self, name: str):
+    def add_single_system(self, name: str) -> None:
         """
         Sync adding a single system.
 
@@ -440,11 +461,11 @@ class CobblerSync:
         if self.settings.manage_dhcp:
             self.dhcp.regen_ethers()
         if self.settings.manage_dns:
-            self.dns.regen_hosts()
+            self.dns.regen_hosts()  # type: ignore
         # write the PXE files for the system
-        self.tftpd.sync_single_system(system)
+        self.tftpd.sync_single_system(system)  # type: ignore
 
-    def remove_single_system(self, name: str):
+    def remove_single_system(self, name: str) -> None:
         """
         Sync removing a single system.
 
@@ -454,6 +475,9 @@ class CobblerSync:
         # delete contents of autoinsts_sys/$name in webdir
         system_record = self.systems.find(name=name)
 
+        if system_record is None or isinstance(system_record, list):
+            raise ValueError("System not found or more then one system found!")
+
         for (interface_name, _) in list(system_record.interfaces.items()):
             pxe_filename = system_record.get_config_filename(
                 interface=interface_name, loader="pxe"
@@ -461,20 +485,22 @@ class CobblerSync:
             grub_filename = system_record.get_config_filename(
                 interface=interface_name, loader="grub"
             )
-            filesystem_helpers.rmfile(
-                os.path.join(bootloc, "pxelinux.cfg", pxe_filename)
-            )
+            if pxe_filename is not None:
+                filesystem_helpers.rmfile(
+                    os.path.join(bootloc, "pxelinux.cfg", pxe_filename)
+                )
             if not (system_record.name == "default" and grub_filename is None):
                 # A default system can't have GRUB entries and thus we want to skip this.
                 filesystem_helpers.rmfile(
-                    os.path.join(bootloc, "grub", "system", grub_filename)
+                    os.path.join(bootloc, "grub", "system", grub_filename)  # type: ignore
                 )
             filesystem_helpers.rmfile(
                 os.path.join(bootloc, "grub", "system_link", system_record.name)
             )
-            filesystem_helpers.rmtree(os.path.join(bootloc, "esxi", pxe_filename))
+            if pxe_filename is not None:
+                filesystem_helpers.rmtree(os.path.join(bootloc, "esxi", pxe_filename))
 
-    def remove_single_menu(self, rebuild_menu: bool = True):
+    def remove_single_menu(self, rebuild_menu: bool = True) -> None:
         """
         Sync removing a single menu.
         :param rebuild_menu: Whether to rebuild the grub/... menu or not.
